Velikiy_K8s_Deploy/app/requirements.txt
Flask==3.0.0
psycopg2-binary==2.9.9
redis==5.0.1
prometheus_flask_exporter==0.23.0
uWSGI==2.0.23